Collect UPnP stuff in one folder
