+++ src/file_request_handler.cc
@@ -45,12 +45,12 @@
 #include "iohandler/file_io_handler.h"
 #include "metadata/metadata_handler.h"
 #include "transcoding/transcode_dispatcher.h"
-#include "upnp_common.h"
-#include "upnp_xml.h"
+#include "upnp/headers.h"
+#include "upnp/quirks.h"
+#include "upnp/upnp_common.h"
+#include "upnp/xml_builder.h"
 #include "util/grb_net.h"
 #include "util/tools.h"
-#include "util/upnp_headers.h"
-#include "util/upnp_quirks.h"
 #include "util/url_utils.h"
 #include "web/session_manager.h"
