+++ src/server.cc
@@ -56,14 +56,15 @@
 #include "file_request_handler.h"
 #include "iohandler/io_handler.h"
 #include "subscription_request.h"
-#include "upnp_cds.h"
-#include "upnp_cm.h"
-#include "upnp_common.h"
-#include "upnp_mrreg.h"
+#include "upnp/clients.h"
+#include "upnp/conn_mgr_service.h"
+#include "upnp/cont_dir_service.h"
+#include "upnp/mr_reg_service.h"
+#include "upnp/upnp_common.h"
+#include "upnp/xml_builder.h"
 #include "util/grb_net.h"
 #include "util/mime.h"
 #include "util/tools.h"
-#include "util/upnp_clients.h"
 #include "util/url_utils.h"
 #include "util/xml_to_json.h"
 #include "web/pages.h"
