+++ src/upnp/clients.cc
@@ -25,7 +25,7 @@
 /// client info initially taken from https://sourceforge.net/p/minidlna/git/ci/master/tree/clients.cc
 #define LOG_FAC log_facility_t::clients
 
-#include "upnp_clients.h" // API
+#include "clients.h" // API
 
 #include "config/config.h"
 #include "config/result/client_config.h"
