+++ test/scripting/test_external_asx_playlist.cc
@@ -28,7 +28,7 @@ Gerbera - https://gerbera.io/
 #include <pugixml.hpp>
 
 #include "cds/cds_objects.h"
-#include "upnp_common.h"
+#include "upnp/upnp_common.h"
 #include "util/string_converter.h"
 #include "util/tools.h"
