+++ test/scripting/test_nfo_metafile.cc
@@ -29,7 +29,7 @@
 #include <pugixml.hpp>
 
 #include "cds/cds_objects.h"
-#include "upnp_common.h"
+#include "upnp/upnp_common.h"
 #include "util/string_converter.h"
 #include "util/tools.h"
